+++ docs/demos/q2_moving_pictures/js/display.js
@@ -142,6 +142,7 @@ define([
                 "colorField",
                 "colorScale",
                 "borderCheckbox",
+                "jitterCheckbox",
             ];
 
             // Set up relevant DOM bindings
@@ -190,6 +191,9 @@ define([
                     },
                     boxplotCheckbox: async function () {
                         await display.updateSamplePlotBoxplot();
+                    },
+                    jitterCheckbox: async function () {
+                        await display.updateSamplePlotJitter();
                     },
                     borderCheckbox: async function () {
                         await display.updateSamplePlotBorders();
@@ -286,8 +290,8 @@ define([
                 // view/select.
                 // TODO: make this a separate func so we can unit-test it
                 if (
-                    this.featureIDs.length <=
-                    this.rankPlotJSON.config.view.width
+                    this.getRankPlotEffectiveNumFeatures() <=
+                    this.getRankPlotFixedWidth()
                 ) {
                     document.getElementById(
                         "fitBarSizeCheckbox"
@@ -511,6 +515,65 @@ define([
             await this.makeRankPlot(true);
         }
 
+        /* Returns a reasonable rank plot width to use for the "constant
+         * plot width" checkbox.
+         *
+         * Previously, we could figure out the width of the rank plot using
+         * this.rankPlotJSON.config.view.width. This is no longer possible
+         * as of 2025. We *can* use .config.view.continuousWidth, but that's
+         * smaller and doesn't seem to actually do anything because the width
+         * of the plot is determined by the step sizes instead. huh. I guess
+         * that makes sense...? It makes you wonder what .config.view.width
+         * even was being set to in the first place.
+         *
+         * So! For the purposes of just, like, setting the bars in the rank
+         * plot to some sort of reasonably small-ish area to make the display
+         * look nice, we can just use jQuery to get a reasonable size for the
+         * rank plot. This is cool because the return value of
+         * $(window).width() scales as we resize the browser window, so
+         * as we call this function it should always give a reasonable-ish rank
+         * plot width. It won't be the SAME fixed width as before, but that
+         * shouldn't be a big deal. The units are the same, anyway (the output
+         * of $(window).width() is pixels).
+         *
+         * See: https://api.jquery.com/width/ and
+         * https://stackoverflow.com/a/1038765/10730311
+         */
+        getRankPlotFixedWidth() {
+            return $(window).width() / 3;
+        }
+
+        /* Returns the outer padding of the rank plot.
+         *
+         * See https://vega.github.io/vega-lite/docs/scale.html#band for a
+         * diagram. Basically, if this is 0, then the leftmost side of the
+         * leftmost feature's bar will be flush with the left side of the plot
+         * (same with the right side); and if this is 1, then we'll have an
+         * extra "bar" worth of space on both sides. And so on.
+         *
+         * This doesn't normally matter, but if there are very few features
+         * (as in e.g. the "matching" integration test) then we want to take
+         * this into account when scaling the plot. The value returned by
+         * this function can be "added on" to the number of features we have
+         * when figuring out how to scale the bar widths.
+         */
+        getRankPlotDoublePaddingOuter() {
+            return this.rankPlotJSON.encoding.x.scale.paddingOuter * 2;
+        }
+
+        getRankPlotEffectiveNumFeatures() {
+            return (
+                this.featureIDs.length + this.getRankPlotDoublePaddingOuter()
+            );
+        }
+
+        getRankPlotFixedWidthBarSize() {
+            return (
+                this.getRankPlotFixedWidth() /
+                this.getRankPlotEffectiveNumFeatures()
+            );
+        }
+
         /* Syncs up the rank plot's bar width with whatever the slider says. */
         async updateRankPlotBarSizeToSlider(callRemakeRankPlot) {
             var sliderBarSize = Number(
@@ -524,12 +587,21 @@ define([
          * Adjusts the "disabled" status of the barSizeSlider accordingly --
          * this prevents users from triggering onchange events while "fitting"
          * the bar widths is enabled.
+         *
+         * At least in firefox, using floating-point steps for bar widths
+         * causes a weird kind of moire effect-looking pattern in the rank
+         * plot. but this (1) has been a thing forever (it is not new i think)
+         * and (2) is probably an artifact of however vega/d3/etc renders stuff
+         * under the hood. not a big deal -- i prefer the flexibility of
+         * allowing users to scrunch up the bar plot as needed, even if the
+         * aforementioned schrunched plots look a little quirky. (we COULD
+         * e.g. round off the widths to fix this, maybe, but then when there
+         * are tons and tons of features we'd either round to 0 [bad] or 1
+         * [too big].
          */
         async updateRankPlotBarFitting(callRemakeRankPlot) {
             if (document.getElementById("fitBarSizeCheckbox").checked) {
-                var fittedBarSize =
-                    this.rankPlotJSON.config.view.width /
-                    this.featureIDs.length;
+                var fittedBarSize = this.getRankPlotFixedWidthBarSize();
                 document.getElementById("barSizeSlider").disabled = true;
                 await this.updateRankPlotBarSize(
                     fittedBarSize,
@@ -548,7 +620,7 @@ define([
          * element invisible.)
          */
         async updateRankPlotBarSize(newBarSize, callRemakeRankPlot) {
-            this.rankPlotJSON.encoding.x.scale.rangeStep = newBarSize;
+            this.rankPlotJSON.width.step = newBarSize;
             if (newBarSize < 1) {
                 document
                     .getElementById("barSizeWarning")
@@ -911,6 +983,11 @@ define([
          * metadata values due to how Qurro's python code processes sample
          * metadata (replacing all NaN values with None, which gets converted
          * to null by json.dumps() in python).
+         *
+         * ... Also, this is unrelated, but while we are initializing the
+         * samplePlotJSON.transform, we also include a transform creating the
+         * "qurro_jitter" field randomly for use with jitter plots. Just in
+         * case the user wants to use those at some point.
          */
         updateSamplePlotFilters() {
             // Figure out the current [x-axis/color] [field/encoding type].
@@ -944,7 +1021,10 @@ define([
                     " && isFinite(toNumber(" + datumColorField + "))";
             }
 
-            this.samplePlotJSON.transform = [{ filter: filterString }];
+            this.samplePlotJSON.transform = [
+                { filter: filterString },
+                { calculate: "random()", as: "qurro_jitter" },
+            ];
         }
 
         /* Update color so that color encoding matches the x-axis encoding
@@ -1110,14 +1190,21 @@ define([
                 // labelAngle parameter.
                 if (newScale === "nominal") {
                     this.samplePlotJSON.encoding.x.axis = { labelAngle: -45 };
+                    // boxplots take precedence over jitter
                     if (document.getElementById("boxplotCheckbox").checked) {
                         this.changeSamplePlotToBoxplot(false);
+                    } else if (
+                        document.getElementById("jitterCheckbox").checked
+                    ) {
+                        this.addSamplePlotJitter();
                     }
                 } else {
                     this.changeSamplePlotFromBoxplot(false);
-                    // This should work even if the axis property is undefined
+                    // Un-rotate the x-axis labels.
+                    // This should work even if the .axis property is undefined
                     // -- it just won't do anything in that case.
                     delete this.samplePlotJSON.encoding.x.axis;
+                    this.removeSamplePlotJitter();
                 }
             } else {
                 this.samplePlotJSON.encoding.color.type = document.getElementById(
@@ -1145,9 +1232,22 @@ define([
         async updateSamplePlotBorders() {
             if (document.getElementById("borderCheckbox").checked) {
                 this.addSamplePlotBorders();
-                await this.remakeSamplePlot();
             } else {
                 this.removeSamplePlotBorders();
+            }
+            await this.remakeSamplePlot();
+        }
+
+        /* Same deal, but for jitter. Can't believe this took my entire phd to
+         * get done lmao
+         */
+        async updateSamplePlotJitter() {
+            if (this.samplePlotJSON.encoding.x.type === "nominal") {
+                if (document.getElementById("jitterCheckbox").checked) {
+                    this.addSamplePlotJitter();
+                } else {
+                    this.removeSamplePlotJitter();
+                }
                 await this.remakeSamplePlot();
             }
         }
@@ -1177,6 +1277,22 @@ define([
             delete this.samplePlotJSON.encoding.color.legend;
         }
 
+        addSamplePlotJitter() {
+            // regarding the scale: qurro_jitter will be a random number
+            // in the range [0, 1]. We can increase padding between adjacent
+            // categories by reducing jitter - and we can do that by just
+            // using a slightly larger domain.
+            this.samplePlotJSON.encoding.xOffset = {
+                field: "qurro_jitter",
+                type: "quantitative",
+                scale: { domain: [-0.25, 1.25] },
+            };
+        }
+
+        removeSamplePlotJitter() {
+            delete this.samplePlotJSON.encoding.xOffset;
+        }
+
         /* Changes the sample plot JSON and DOM elements to get ready for
          * switching to "boxplot mode." If callRemakeSamplePlot is truthy, this
          * will actually call this.remakeSamplePlot(); otherwise, this won't do
@@ -1194,6 +1310,7 @@ define([
          */
         async changeSamplePlotToBoxplot(callRemakeSamplePlot) {
             this.removeSamplePlotBorders();
+            this.removeSamplePlotJitter();
             this.samplePlotJSON.mark.type = "boxplot";
             // Make the middle tick of the boxplot black. This makes boxes for
             // which only one sample is available show up on the white
@@ -1216,6 +1333,9 @@ define([
         async changeSamplePlotFromBoxplot(callRemakeSamplePlot) {
             if (document.getElementById("borderCheckbox").checked) {
                 this.addSamplePlotBorders();
+            }
+            if (document.getElementById("jitterCheckbox").checked) {
+                this.addSamplePlotJitter();
             }
             this.samplePlotJSON.mark.type = "circle";
             delete this.samplePlotJSON.mark.median;
@@ -1535,6 +1655,8 @@ define([
                 document.getElementById("boxplotCheckbox").checked = false;
                 // ... And the sample border checkbox
                 document.getElementById("borderCheckbox").checked = false;
+                // ... And the jitter checkbox
+                document.getElementById("jitterCheckbox").checked = false;
                 // ... And the "exclude metadata fields" checkbox
                 document.getElementById(
                     "exclSMFieldsInExportCheckbox"